--- src/components/Adsense/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Adsense';
