+++ src/containers/Respect/Main.tsx
@@ -1,4 +1,5 @@
 import React, { FC } from 'react';
+import AppAdsense from '@/components/AppAdsense';
 import CardView from '@/containers/Respect/CardView';
 import MainTemplate from '@/templates/MainTemplate';
 
@@ -10,6 +11,7 @@ const Main: FC<MainProps> = ({ data }) => {
   return (
     <MainTemplate label='respect-main'>
       <CardView data={data} />
+      <AppAdsense />
     </MainTemplate>
   );
 };
